fix: consolidate new DocTR tests to minimize ONNX sessions on Windows
The Windows CI runner accumulates unreleased BFC arena memory across
ONNX sessions (each ONNXOcr load/unload cycle). Pre-PR the suite had
~16 sessions and passed. The new test files added 10 more, pushing
the total to 26 and causing OOM failures.

Restructure to add only 1 new session:

- doctr-pipeline.test.js: ONE shared loaded mobilenet instance for all
  tests (sequential runs, coordinate scaling, unrecognizable text,
  run-after-unload, double-unload). Was 3 separate files with 10 sessions.
- doctr-param-validation.test.js: add run-before-load test (0 sessions,
  never calls load()).
- Delete doctr-scaling.test.js (assertions moved to pipeline).
- Delete doctr-session-lifecycle.test.js (edge cases moved to pipeline
  teardown and param-validation; reload/cancel/batchSize already covered
  by EasyOCR lifecycle.test.js and doctr-models.test.js).

New total: ~17 sessions (was 26, pre-PR was 16).

Made-with: Cursor
